[#2029] Changed the drag and drop behavior to not activate when clicking links.
When a drag and drop action was started by clicking a link, the browser would
interpret the drop event as a click, which would send the user to the link
instead of doing the drag and drop.  This displayed an error message since the
Ajax request didn't complete and sent the user to the target page.  Disabling
the click event wouldn't work without removing the links so I just disallowed
the drag and drop to start on a link.