Some changes made most of this plugin work on redmine 2.3.3
First and most impoertant the directory name was changed to
"stuff_to_do_plugin" because it has to fit the plugin id defined in
init.rb. Otherwise rails won't find it.
Additionally I made some minor changes which were necessary to be
compatible with the new rails and redmine versions.
Last but not least I added translations for the permissions added in
german and english.